+++ apps/runtime-demo/3005-runtime-host/remotes.d.ts
@@ -1,4 +1,4 @@
-declare module 'remote1/useCustomRemoteHook';
-declare module 'remote1/WebpackSvg';
-declare module 'remote1/WebpackPng';
-declare module 'dynamic-remote/ButtonOldAnt';
+// declare module 'remote1/useCustomRemoteHook';
+// declare module 'remote1/WebpackSvg';
+// declare module 'remote1/WebpackPng';
+// declare module 'dynamic-remote/ButtonOldAnt';