+++ apps/runtime-demo/3005-runtime-host/src/bootstrap.tsx
@@ -1,20 +1,20 @@
 import React, { StrictMode } from 'react';
-import { init } from '@module-federation/runtime';
-import customPlugin from './runtimePlugin';
-
+import {
+  init,
+  registerGlobalPlugins,
+} from '@module-federation/enhanced/runtime';
 import * as ReactDOM from 'react-dom/client';
-
 import App from './App';
+
 init({
   name: 'runtime_host',
   remotes: [
     {
       name: 'runtime_remote2',
       alias: 'dynamic-remote',
-      entry: 'http://localhost:3007/remoteEntry.js',
+      entry: 'http://localhost:3007/mf-manifest.json',
     },
   ],
-  plugins: [customPlugin()],
 });
 
 const root = ReactDOM.createRoot(