+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @module-federation/modernjs
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [8f3a440]
+- Updated dependencies [2f697b9]
+- Updated dependencies [8f3a440]
+  - @module-federation/enhanced@0.1.4
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [32eba3c]
+- Updated dependencies [6b3b210]
+  - @module-federation/enhanced@0.1.3
+
 ## 0.1.7
 
 ### Patch Changes